% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetbox.R
\name{controlChartDygraph}
\alias{controlChartDygraph}
\title{Ploting Dygraph in R}
\usage{
controlChartDygraph(tsData, headerText, enableDyrange = FALSE,
  enableDyhighlight = FALSE, legendalignment = "centre",
  yaxisHeader = NULL)
}
\arguments{
\item{tsData}{It will contain either the xts object or a numeric data frame.}

\item{headerText}{It will show the Header Text Explaining what the Dygraph Plot is about.}

\item{enableDyrange}{If \code{TRUE}, It will display the DyRange Selector Under the Dygraph.}

\item{enableDyhighlight}{if \code{TRUE}, It will Display the series as highlighted whenever hovered over.}

\item{legendalignment}{This will control the Text Alignment for showing Dygraph Plot values over dygraph chart.}

\item{yaxisHeader}{It will show the Text for the Y-axis Plot.}
}
\description{
Function to render a Dygraph Widget that plot time series data over customisable axes.
}
