% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetbox.R
\name{highChart}
\alias{highChart}
\title{Plotting Highcharter Control Charts in R}
\usage{
highChart(title, theme = "sandsignika", seriesData = NULL,
  seriesCategories = NULL)
}
\arguments{
\item{title}{Title for the Highcharts.}

\item{theme}{Highchart theme to be used in the Highchart Plot.}

\item{seriesData}{Series to be plotted into the HighChart.}

\item{seriesCategories}{The X-axis categories to plot on Highchart.}
}
\description{
Functions to render a Plot of Highcharter's Library
}
