% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgetbox.R
\name{multilineChart}
\alias{multilineChart}
\title{Plotting Multi-line charts in R}
\usage{
multilineChart(multilinedata)
}
\arguments{
\item{multilinedata}{It will contain the whole dataframe passed by the user to the function.}
}
\description{
Functions to render a plot of Multi-line chart of Metrics Graphics Library.
}
